+++ internal/infrastructure/db/postgres/nndn.go
@@ -109,7 +109,7 @@ func (r *GormNNDNRepository) DeleteNNDN(ctx context.Context, name string) error 
 	return result.Error
 }
 
-func (r *GormNNDNRepository) Count(ctx context.Context, filter queries.ListNndnsFilter) (int, error) {
+func (r *GormNNDNRepository) Count(ctx context.Context, filter queries.ListNndnsFilter) (int64, error) {
 	dbQuery := r.db.WithContext(ctx).Model(&NNDN{})
 	dbQuery, err := setNNDNFilters(dbQuery, filter)
 	if err != nil {
@@ -117,7 +117,7 @@ func (r *GormNNDNRepository) Count(ctx context.Context, filter queries.ListNndns
 	}
 	var count int64
 	err = dbQuery.Count(&count).Error
-	return int(count), err
+	return count, err
 }
 
 func (r *GormNNDNRepository) ListNNDNs(ctx context.Context, params queries.ListItemsQuery) ([]*entities.NNDN, string, error) {
